+++ packages/omen/lib/omen.dart
@@ -1,3 +1,3 @@
-export 'src/code_unit_distribution.dart';
+export 'src/dice.dart';
 export 'src/distribution.dart';
-export 'src/list_distribution.dart';
+export 'src/shuffler.dart';