+++ packages/omen/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- #endregion -->
 
+## 0.1.1
+
+**New features**:
+
+- Added `Dice` class (implements `Distribution<int>`).
+- Added `Shuffler` interface, with default implementations.
+
 ## 0.1.0-alpha
 
 Initial pre-release of the `omen` package.